dictionary additions; formatting; field annotation
Add some words to the team-shared ReSharper dictionary.

Format `MainWindow.axaml` with "XAML Styler" VS extension.

Create `_heading` field annotated with `ObservableProperty` attribute in
`MainWindowViewModel`.

Annotating fields in this way allows us to "run additional logic when the
property is about to be updated and right after it is updated, if needed."

See
https://learn.microsoft.com/en-us/dotnet/communitytoolkit/mvvm/generators/observableproperty
for more details.